inserted in modelCreating relationship between tables in ef, enchanted the structures of models, add new logics based on new structure.

+++ MAServices/Interfaces/IReviewServices.cs
@@ -4,7 +4,7 @@ namespace MAServices.Interfaces
 {
     public interface IReviewServices
     {
-        Task PostNewReview(User user, Movie movie, string? descriptionVote, int vote);
+        Task PostNewReview(User user, Movie movie, string? descriptionVote, short vote);
 
         Task<ICollection<Review>> GetReviews();
